added repository for grades(SubjectDetails) added Update functions for mainForm added YearListItem class to be used in Years dropdown list to store year_ids Working adding and editing the year NOTE: cascading removing for DB not enabled - fix this or workaround

+++ GradingBookProject/Data/YearsRepository.cs
@@ -15,10 +15,6 @@ namespace GradingBookProject.Data
             context = new GradingBookDbEntities();
         }
 
-        /// <summary>
-        /// adds a Year to a database.
-        /// </summary>
-        /// <param name="year">Year to be added to a DB</param>
         public void AddYear(Years year, int userid)
         {
             if (context.Users.FirstOrDefault(u => u.id == userid).Years.FirstOrDefault(y => y.id == year.id) != null)
@@ -28,32 +24,43 @@ namespace GradingBookProject.Data
             context.SaveChanges();
         }
 
-        /// <summary>
-        /// return Years as a list (IEnumerable)
-        /// </summary>
         public IEnumerable<Years> Years(int userid)
         {
+            if (context.Years.FirstOrDefault(u=> u.id == userid) == null)
+                throw new Exception("User has no Years.");
+
             return context.Users.FirstOrDefault(u => u.id == userid).Years;
+        }
+
+        public Years Year(int yearid, int userid)
+        {
+            if (context.Users.FirstOrDefault(u => u.id == userid).Years.FirstOrDefault(y => y.id == yearid) == null)
+                throw new Exception("Such year does not exist!");
+
+            return context.Users.FirstOrDefault(u => u.id == userid).Years.FirstOrDefault(y => y.id == yearid); 
         }
 
         public void UpdateYear(Years year)
         {
+            if (context.Years.FirstOrDefault(y => y.id == year.id) == null)
+                throw new Exception("Such year does not exist!");
+
             context.Years.FirstOrDefault(y => y.id == year.id).name = year.name;
             context.Years.FirstOrDefault(y => y.id == year.id).start = year.start;
             context.Years.FirstOrDefault(y => y.id == year.id).end_date = year.end_date;
             context.Years.FirstOrDefault(y => y.id == year.id).year_desc = year.year_desc;
             context.SaveChanges();
-        }
+        }   
 
-        /// <summary>
-        /// Deletes a given Year from a database (cascading).
-        /// </summary>
-        /// <param name="year">Year to be deleted from a DB</param>
         public void DeleteYear(Years year, int userid) 
         {
+            if (context.Users.FirstOrDefault(u => u.id == userid).Years.FirstOrDefault(y => y.id == year.id) == null)
+                throw new Exception("Such year doesn't exist!");
+
             context.Years.Remove(context.Users.FirstOrDefault(u => u.id == userid).Years.FirstOrDefault(y => y.id == year.id));
             context.SaveChanges();
         }
+
 
     }
 }
